+++ src/components/ListView/BsListTile.ts
@@ -1,11 +1,36 @@
 import type { ComponentOptionsMixin, ComputedOptions, EmitsOptions, MethodOptions } from 'vue';
-import { computed, defineComponent, inject, nextTick, onBeforeUpdate, onMounted, ref, shallowRef, watch } from 'vue';
-import { useCurrentRoute, useHasLink, useHasRouter } from '../../mixins/CommonApi';
-import type { IListItem, IListViewProvider, TBsListTile, TListTileOptionProps, TRecord } from '../../types';
+import {
+    computed,
+    defineComponent,
+    inject,
+    nextTick,
+    onBeforeUpdate,
+    onMounted,
+    ref,
+    shallowRef,
+    watch,
+} from 'vue';
+import { useCurrentRoute, useHasLink, useHasRouter, useIsRouteMatch } from '../../mixins/CommonApi';
+import type {
+    IListItem,
+    IListViewProvider,
+    TBsListTile,
+    TListTileOptionProps,
+    TRecord,
+} from '../../types';
 import { useListTileClassNames, useRenderListTile } from './mixins/listTileApi';
 import { listTileProps } from './mixins/listViewProps';
 
-export default defineComponent<TBsListTile, TRecord, TRecord, ComputedOptions, MethodOptions, ComponentOptionsMixin, ComponentOptionsMixin, EmitsOptions>({
+export default defineComponent<
+    TBsListTile,
+    TRecord,
+    TRecord,
+    ComputedOptions,
+    MethodOptions,
+    ComponentOptionsMixin,
+    ComponentOptionsMixin,
+    EmitsOptions
+>({
     name: 'BsListTile',
     props: listTileProps,
     emits: [
@@ -13,52 +38,48 @@ export default defineComponent<TBsListTile, TRecord, TRecord, ComputedOptions, M
         /**
          * Fired when this component's state is updated.
          */
-        'update:active'
+        'update:active',
     ],
-    setup(props, {emit, expose, slots}) {
+    setup(props, { emit, expose, slots }) {
         const thisProps = props as Readonly<TListTileOptionProps>;
         const vm = shallowRef<IListItem>();
         const isActive = ref<boolean | undefined>(thisProps.active);
         const hasRouter = ref<boolean>(useHasRouter(thisProps));
         const hasLink = ref<boolean>(useHasLink(thisProps));
 
-        expose({isActive});
+        expose({ isActive });
 
         const provider = inject<IListViewProvider>('ListView');
-        const tagName = computed<string>(
-            () => hasRouter.value || hasLink.value ? 'a' : 'div'
-        );
-        const tileClasses = computed<TRecord>(
-            () => useListTileClassNames(tagName.value, thisProps, isActive, hasLink, provider)
+        const tagName = computed<string>(() => (hasRouter.value || hasLink.value ? 'a' : 'div'));
+        const tileClasses = computed<TRecord>(() =>
+            useListTileClassNames(tagName.value, thisProps, isActive, hasLink, provider)
         );
 
         watch(
             () => thisProps.active,
             (value) => {
                 if (provider && provider.config.individualState === true) {
-                    isActive.value = value
+                    isActive.value = value;
                 }
             }
         );
-        onBeforeUpdate(
-            () => {
-                if (hasRouter.value) {
-                    const route = useCurrentRoute();
-                    if (route && route.value.path === thisProps.path) {
-                        isActive.value = true;
-                        nextTick().then(() => {
-                            emit('update:active', true);
-                        });
-                    }
+        onBeforeUpdate(() => {
+            if (hasRouter.value) {
+                const route = useCurrentRoute();
+                if (route && useIsRouteMatch(route, thisProps)) {
+                    isActive.value = true;
+                    nextTick().then(() => {
+                        emit('update:active', true);
+                    });
                 }
             }
-        );
+        });
         onMounted(() => {
             hasRouter.value = useHasRouter(thisProps);
             hasLink.value = useHasLink(thisProps);
         });
 
         return () =>
-            useRenderListTile(tagName.value, slots, emit, thisProps, tileClasses, vm, provider)
-    }
+            useRenderListTile(tagName.value, slots, emit, thisProps, tileClasses, vm, provider);
+    },
 });