+++ src/components/ListView/mixins/listTileApi.ts
@@ -8,37 +8,38 @@ import type {
     TBsRipple,
     TEmitFn,
     TListTileOptionProps,
-    TRecord
+    TRecord,
 } from '../../../types';
 import { BsRipple } from '../../Animation';
 import ListItem from './ListItem';
-
 
 export function useListTileClassNames(
     tagName: string,
     props: Readonly<TListTileOptionProps>,
     isActive: Ref<boolean | undefined>,
     hasLink: Ref<boolean>,
-    provider?: IListViewProvider,
+    provider?: IListViewProvider
 ): TRecord {
     return {
         [`${cssPrefix}list-tile d-flex`]: true,
         [`${cssPrefix}link`]: (tagName === 'a' || props.navigable) && !props.disabled,
-        [`${cssPrefix}tile-border-${provider?.itemBorderVariant}`]: (
-            provider?.itemBorderVariant && !props.borderOff &&
-            ['left', 'right', 'left-right', 'top', 'bottom', 'top-bottom'].includes(provider.itemBorderVariant)
-        ),
-        [`${cssPrefix}tile-space-${provider?.spaceAround}`]: (
-            provider?.spaceAround && ['both', 'left', 'right'].includes(provider.spaceAround)
-        ),
-        [`${props.activeClass}`]: (
+        [`${cssPrefix}tile-border-${provider?.itemBorderVariant}`]:
+            provider?.itemBorderVariant &&
+            !props.borderOff &&
+            ['left', 'right', 'left-right', 'top', 'bottom', 'top-bottom'].includes(
+                provider.itemBorderVariant
+            ),
+        [`${cssPrefix}tile-space-${provider?.spaceAround}`]:
+            provider?.spaceAround && ['both', 'left', 'right'].includes(provider.spaceAround),
+        [`${props.activeClass}`]:
             (hasLink.value || props.navigable) &&
-            props.activeClass && !props.disabled && isActive.value === true
-        ),
-        'active': (hasLink.value || props.navigable) && !props.disabled && isActive.value === true,
-        'rounded': provider?.itemRounded === true && !props.roundedOff,
+            props.activeClass &&
+            !props.disabled &&
+            isActive.value === true,
+        active: (hasLink.value || props.navigable) && !props.disabled && isActive.value === true,
+        rounded: provider?.itemRounded === true && !props.roundedOff,
         'rounded-pill': provider?.itemRoundedPill === true && !props.pillOff,
-        'disabled': props.disabled === true,
+        disabled: props.disabled === true,
     };
 }
 
@@ -49,42 +50,57 @@ function createListTileElement(
     props: Readonly<TListTileOptionProps>,
     classes: ComputedRef<TRecord>,
     instance: ShallowRef<IListItem | undefined>,
-    provider?: IListViewProvider,
+    provider?: IListViewProvider
 ): VNode {
-    return h(tagName, {
-        id: props.id,
-        class: classes.value,
-        href: tagName === 'a' ? props.url : undefined,
-        onVnodeMounted: (vNode: VNode) => {
-            instance.value = new ListItem(<string>props.id, 'BsListTile', (<IVNode>vNode).ctx, emit);
-            provider?.addItem(instance.value);
-        },
-        onVnodeBeforeUnmount: () => instance.value && provider?.removeItem(instance.value),
-        onClick: (e: Event) => {
-            // console.log("tile-instance", instance.value);
-            if (tagName === 'a' || props.navigable) {
-                if (provider) {
-                    provider.activeItem = instance.value;
+    return h(
+        tagName,
+        {
+            id: props.id,
+            class: classes.value,
+            href: tagName === 'a' ? props.url : undefined,
+            onVnodeMounted: (vNode: VNode) => {
+                instance.value = new ListItem(
+                    <string>props.id,
+                    'BsListTile',
+                    (<IVNode>vNode).ctx,
+                    emit
+                );
+                provider?.addItem(instance.value);
+            },
+            onVnodeBeforeUnmount: () => instance.value && provider?.removeItem(instance.value),
+            onClick: (e: Event) => {
+                // console.log("tile-instance", instance.value);
+                if (tagName === 'a' || props.navigable) {
+                    if (provider) {
+                        provider.activeItem = instance.value;
+                    }
+                    if (!provider || provider.config.individualState === true) {
+                        emit('update:active', !props.active);
+                    }
                 }
-                if (!provider || provider.config.individualState === true) {
-                    emit('update:active', !props.active);
-                }
-            }
-            emit('click', e, (instance.value ? instance.value.component.vnode.el : undefined));
+                emit('click', e, instance.value ? instance.value.component.vnode.el : undefined);
+            },
         },
-    }, [
-        h<TBsRipple>(BsRipple, {
-            class: [
-                'd-flex',
-                provider?.itemRounded === true && !props.roundedOff ? 'rounded' : '',
-                provider?.itemRoundedPill === true && !props.pillOff ? 'rounded-pill' : '',
-            ],
-            // @ts-ignore
-            disabled: (props.rippleOff || props.disabled || !(tagName === 'a' || props.navigable)) as Prop<boolean>,
-        }, {
-            default: () => slots.default && slots.default()
-        })
-    ]);
+        [
+            h<TBsRipple>(
+                BsRipple,
+                {
+                    class: [
+                        'd-flex',
+                        provider?.itemRounded === true && !props.roundedOff ? 'rounded' : '',
+                        provider?.itemRoundedPill === true && !props.pillOff ? 'rounded-pill' : '',
+                    ],
+                    // @ts-ignore
+                    disabled: (props.rippleOff ||
+                        props.disabled ||
+                        !(tagName === 'a' || props.navigable)) as Prop<boolean>,
+                },
+                {
+                    default: () => slots.default && slots.default(),
+                }
+            ),
+        ]
+    );
 }
 
 function createListTileRouterElement(
@@ -93,15 +109,21 @@ function createListTileRouterElement(
     props: Readonly<TListTileOptionProps>,
     classes: ComputedRef<TRecord>,
     instance: ShallowRef<IListItem | undefined>,
-    provider?: IListViewProvider,
+    provider?: IListViewProvider
 ): VNode {
-    return useRenderRouter({
+    return useRenderRouter(
+        {
             id: props.id,
             class: classes.value,
             activeClass: props.activeClass || 'active',
-            to: props.path,
+            to: props.location ?? (props.pathName ? { name: props.pathName } : props.path),
             onVnodeMounted: (vNode: VNode) => {
-                instance.value = new ListItem(<string>props.id, 'BsListTile', (vNode as IVNode).ctx, emit);
+                instance.value = new ListItem(
+                    props.id as string,
+                    'BsListTile',
+                    (vNode as IVNode).ctx,
+                    emit
+                );
                 provider?.addItem(instance.value);
             },
             onVnodeBeforeUnmount: () => instance.value && provider?.removeItem(instance.value),
@@ -113,20 +135,24 @@ function createListTileRouterElement(
                     emit('update:active', !props.active);
                 }
                 emit('click', e, instance.value?.component.vnode.el);
-            }
+            },
         },
         [
-            h<TBsRipple>(BsRipple, {
-                class: [
-                    'd-flex',
-                    provider?.itemRounded === true && !props.roundedOff ? 'rounded' : '',
-                    provider?.itemRoundedPill === true && !props.pillOff ? 'rounded-pill' : '',
-                ],
-                // @ts-ignore
-                disabled: (props.rippleOff || props.disabled) as Prop<boolean>,
-            }, {
-                default: () => slots.default && slots.default()
-            })
+            h<TBsRipple>(
+                BsRipple,
+                {
+                    class: [
+                        'd-flex',
+                        provider?.itemRounded === true && !props.roundedOff ? 'rounded' : '',
+                        provider?.itemRoundedPill === true && !props.pillOff ? 'rounded-pill' : '',
+                    ],
+                    // @ts-ignore
+                    disabled: (props.rippleOff || props.disabled) as Prop<boolean>,
+                },
+                {
+                    default: () => slots.default && slots.default(),
+                }
+            ),
         ]
     );
 }
@@ -138,7 +164,7 @@ export function useRenderListTile(
     props: Readonly<TListTileOptionProps>,
     classes: ComputedRef<TRecord>,
     instance: ShallowRef<IListItem | undefined>,
-    provider?: IListViewProvider,
+    provider?: IListViewProvider
 ): VNode {
     if (useHasRouter(props) && !props.disabled) {
         return createListTileRouterElement(slots, emit, props, classes, instance, provider);